requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
apscheduler==3.11.0
sqlalchemy
pydantic==2.11.9
python-multipart
passlib
psycopg2-binary==2.9.9
python-dotenv
jose
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing-extensions==4.15.0
tzlocal==5.3.1
pymysql
